refactor: extract RAG components into internal/rag package
