+++ internal/llm/parser_security_test.go
@@ -56,7 +56,7 @@ func TestExtractTag_Security(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, ok := extractTag(tt.content, tt.tag)
+			got, ok := ExtractTag(tt.content, tt.tag)
 			assert.Equal(t, tt.ok, ok)
 			if ok {
 				assert.Equal(t, tt.expected, got)
