+++ internal/rag/arch_context_test.go
@@ -1,4 +1,4 @@
-package llm
+package rag
 
 import (
 	"testing"
