+++ internal/rag/rag_impact.go
@@ -1,4 +1,4 @@
-package llm
+package rag
 
 import (
 	"context"
