+++ internal/rag/rag_rerank_test.go
@@ -1,4 +1,4 @@
-package llm
+package rag
 
 import (
 	"context"
