+++ internal/rag/rag_test.go
@@ -1,4 +1,4 @@
-package llm
+package rag
 
 import (
 	"strings"
